chore. filter 및 member entity 수정

+++ src/main/java/com/sparta/deliveryapp/jwt/JwtUtil.java
@@ -40,9 +40,8 @@ public class JwtUtil {
     }
 
     //토큰 생성
-    public String createToken(Integer userId, String username, UserRole userRole, boolean isActive, boolean isSecession) {
+    public String createToken(Long userId, String username, UserRole userRole, boolean isActive, boolean isSecession) {
         Date date = new Date();
-
         return BEARER_PREFIX +
                 Jwts.builder()
                         .setSubject(String.valueOf(userId))
